Draw multiple borders if a layout's bordercolors is a list
This commit allows passing a list of colours to a layout, and it will
pass those to the window for painting of multiple, different coloured
borders. It also updates the docstrings of the layouts from "border
color" to "border color(s)" and uncomments the Bsp layout using multiple
borders in the default config, both so that people are aware of the new
feature. Any border width can be used, and border painting will try to
evenly distribute pixels across the colours. A test is added.