Set up notifications once upon import and implement notification IDs
A bool flag is used to track whether notifications dependencies are
available, and send_notification will silently do nothing if they are
not. Notification IDs can be passed as a kwarg (id=4), which the
notification service can use to replace existing notifications with the
same ID. The ID is returned, so None can be passed initially and a new
ID will be created and returned for further use.